# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sistem Manajemen Karyawan API",
        "description": "Dokumentasi API untuk sistem manajemen karyawan, absensi, dan penggajian",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "API Server"
        }
    ],
    "paths": {
        "/absensi": {
            "get": {
                "tags": [
                    "Absensi"
                ],
                "summary": "Mendapatkan semua data absensi",
                "operationId": "0b6025737b6ed1ec2bd3573225bb1d08",
                "responses": {
                    "200": {
                        "description": "Berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Absensi retrieved successfully."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Absensi"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Absensi"
                ],
                "summary": "Membuat data absensi baru",
                "operationId": "de8b1bca43458a5c859b97c8f98d71f2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id_absensi": {
                                        "type": "string",
                                        "example": "ABS-001"
                                    },
                                    "id_karyawan": {
                                        "type": "string",
                                        "example": "KRY-001"
                                    },
                                    "waktu": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-04-23 08:00:00"
                                    },
                                    "keterangan": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 201
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Absensi created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Absensi"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                }
            }
        },
        "/absensi/{id_absensi}": {
            "get": {
                "tags": [
                    "Absensi"
                ],
                "summary": "Mendapatkan data absensi berdasarkan ID",
                "operationId": "1db54ba680fc4d08345e1b5bc392be12",
                "parameters": [
                    {
                        "name": "id_absensi",
                        "in": "path",
                        "description": "ID absensi",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Absensi retrieved successfully."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Absensi"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Absensi not found."
                                        },
                                        "data": {
                                            "type": "null",
                                            "example": null
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Absensi"
                ],
                "summary": "Memperbarui data absensi",
                "operationId": "5baaa5647b4ab4d91e09eaa03d80c9f3",
                "parameters": [
                    {
                        "name": "id_absensi",
                        "in": "path",
                        "description": "ID absensi",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id_karyawan": {
                                        "type": "string",
                                        "example": "KRY-002"
                                    },
                                    "waktu": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-04-23 09:30:00"
                                    },
                                    "keterangan": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Berhasil diperbarui",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Absensi updated successfully."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Absensi"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Absensi not found."
                                        },
                                        "data": {
                                            "type": "null",
                                            "example": null
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Absensi"
                ],
                "summary": "Menghapus data absensi",
                "operationId": "361370e575f472d04baee4a4af2a8e8e",
                "parameters": [
                    {
                        "name": "id_absensi",
                        "in": "path",
                        "description": "ID absensi",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil dihapus",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Absensi deleted successfully."
                                        },
                                        "data": {
                                            "type": "null",
                                            "example": null
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Absensi not found."
                                        },
                                        "data": {
                                            "type": "null",
                                            "example": null
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Absensi": {
                "properties": {
                    "id_absensi": {
                        "type": "string",
                        "example": "ABS-001"
                    },
                    "id_karyawan": {
                        "type": "string",
                        "example": "KRY-001"
                    },
                    "waktu": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-04-23 08:00:00"
                    },
                    "keterangan": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "description": "Masukkan API key dengan format Bearer {token}",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}